+++ site/guidelines/displaying-data/index.jsx
@@ -15,7 +15,7 @@ import CodeBlock from 'app_modules/ui/code-block';
 import { prefix as pf } from 'app_modules/ui/util/component';
 
 export default (
-  <PageBody anchorTitle="Displaying Data" contentClassName={pf('site-design site-design-layout')}>
+  <PageBody anchorTitle="Displaying Data" contentClassName={pf('site-guidelines site-guidelines-layout')}>
     <div className={pf('container--large')}>
       <p className="site-text-introduction">Lists and feeds help users navigate data either by theme or chronology. These various ways of displaying data are optimized to help users locate, browse, and work together on records.</p>