+++ site/guidelines/layout/index.jsx
@@ -15,7 +15,7 @@ import CodeBlock from 'app_modules/ui/code-block';
 import { prefix as pf } from 'app_modules/ui/util/component';
 
 export default (
-  <PageBody anchorTitle="Layout" contentClassName={pf('site-design site-design-layout')}>
+  <PageBody anchorTitle="Layout" contentClassName={pf('site-guidelines site-guidelines-layout')}>
     <div className={pf('container--large')}>
       <p className="site-text-introduction">A layout is a structural template that supports consistency across applications by defining visual grids, spacing, and sections. </p>
       <p>Before selecting a layout, review these guidelines:</p>