+++ site/guidelines/navigation/index.jsx
@@ -15,7 +15,7 @@ import CodeBlock from 'app_modules/ui/code-block';
 import { prefix as pf } from 'app_modules/ui/util/component';
 
 export default (
-  <PageBody anchorTitle="Navigation" contentClassName={pf('site-design site-design-navigation')}>
+  <PageBody anchorTitle="Navigation" contentClassName={pf('site-guidelines site-guidelines-navigation')}>
     <div className={pf('container--large')}>
       <p className="site-text-introduction">Navigational patterns help users move between pages and situate themselves in the application.</p>
       <h2 className="site-text-heading--large">Tabs</h2>