+++ site/guidelines/typography/index.jsx
@@ -15,7 +15,7 @@ import PageBody from 'app_modules/site/components/page/body';
 import { prefix as pf } from 'app_modules/ui/util/component';
 
 export default (
-  <PageBody anchorTitle="Typography" contentClassName={pf('site-design site-design-layout')}>
+  <PageBody anchorTitle="Typography" contentClassName={pf('site-guidelines site-guidelines-layout')}>
     <div className={pf('container--large')}>
       <p className="site-text-introduction">Text is the primary way that users digest content and accomplish work, so it’s important to use good typographic principles to establish a clear visual hierarchy and to maximize legibility.</p>
       <h2 className="site-text-heading--large">Salesforce Sans</h2>